Refactoring JWT implementation
